feat(support): move support link from avatar dropdown to global footer
The "Contact support" item was buried inside the user-avatar dropdown
where (1) logged-in users had to discover it by clicking the avatar
and (2) accidental clicks while aiming for Logout fired a mailto: that
opened Mac Mail unwantedly.

Move it to the global VersionFooter so it's always visible without
clicks, and remove it from AuthNav. Footer line reads "Need help?
Contact support" next to the copyright; mailto: still pre-fills with
the account's email + display name when the user is logged in (and
falls through to a generic "Help request" subject for anonymous
visitors).

Refs: SB-35
